コンポーネントにして分けました
