+++ src/components/login/Mypage.jsx
@@ -1,13 +1,14 @@
 // 各必要な要素を取得
-import { onAuthStateChanged, signOut } from "firebase/auth";
+import { signOut } from "firebase/auth";
 import React, { useContext, useEffect, useState } from "react";
 import { useNavigate, Link, useParams } from "react-router-dom";
 import "../../css/components/Mypage.css";
 import { auth } from "../../utils/firebase";
 import AppContext from "../../context/AppContext";
-import { collection, query, where, onSnapshot, serverTimestamp, doc, addDoc, getDocs, deleteDoc } from "firebase/firestore";
+import { collection, query, where, onSnapshot} from "firebase/firestore";
 import { db } from "../../utils/firebase";
 import GetUserInfo from "./GetUserInfo";
+import { Follow } from "./Follow";
 
 //ログイン情報の取得
 const Mypage = () => {
@@ -16,7 +17,7 @@ const Mypage = () => {
   const { user, loading } = useContext(AppContext);
   // Firebaseから取得した内容をpostsに代入
   const [posts, setPosts] = useState([]);
-  const [isFollowing, setIsFollowing] = useState();
+ 
 
   useEffect(() => {
     console.log(user);
@@ -56,56 +57,7 @@ const Mypage = () => {
     };
   }, [user,id]);
 
-  // 現在訪れているユーザーをフォローしているかチェック
-  useEffect(() => {
-    const checkFollowing = async () => {
-        const followingRef = collection(db, "users", user.uid, "following");
-        const q = query(followingRef, where("followingId", "==", id));
-        const querySnapshot = await getDocs(q);
-        if (!querySnapshot.empty) {
-            setIsFollowing(true);
-        } else {
-            setIsFollowing(false);
-        }
-    };
-
-    checkFollowing();
-}, [user, id, isFollowing]);
-
-// フォローボタンの処理
-  const handleFollow = async(e) => {
-    e.preventDefault();
-    
-    const followingRef = collection(db, "users", user.uid, "following");
-    await addDoc(followingRef, {
-      followingId: id,
-      followedAt: serverTimestamp()
-    });
-    
-    setIsFollowing(true);
-    alert("フォローしました！");
-  }
-
-  // フォロー解除ボタンの処理
-  const handleUnfollow = async(e) => {
-    e.preventDefault();
-// まずはクエリで、現在見ているユーザーページのパラムを取得し、それをもとにフォローしているユーザーのドキュメントを取得
-    const followingRef = collection(db, "users", user.uid, "following");
-    const q = query(followingRef, where("followingId", "==", id));
-    const querySnapshot = await getDocs(q);
-// ユーザーが存在していたら、そのドキュメントを削除
-if (!querySnapshot.empty) {
-  querySnapshot.forEach(async (docSnapshot) => {
-        console.log(docSnapshot.id)
-          const docRef = doc(db, "users", user.uid, "following", docSnapshot.id);
-          await deleteDoc(docRef);
-      });
-    }else{
-
-    }
-    setIsFollowing(false)
-    alert("フォロー解除しました！")
-  }
+  
 
   return (
     <div className="mypage">
@@ -121,13 +73,7 @@ if (!querySnapshot.empty) {
         :
         ""
       }
-{/* 現在訪れているユーザーをフォローしているかどうかでボタンが変わる */}
-      {isFollowing 
-        ? 
-        <button onClick={handleUnfollow}>フォロー解除</button>
-        :
-        <button onClick={handleFollow}>フォローする</button>
-    }
+      <Follow/>
       <div className="mypage-list">
         <h2>ブログ一覧</h2>
         <div className="mypage-list__items">
